requirements.txt
pandas==2.1.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
instabot==0.117.0
moviepy==1.0.3
Pillow==10.0.0
nltk==3.8.1
flask==2.3.3
werkzeug==2.3.7
numpy==1.24.3
opencv-python==4.8.1.78 